requirements.txt
streamlit==1.15.0
tensorflow==2.12.0
numpy==1.23.5
Pillow==9.3.0
matplotlib==3.6.2
